+++ lib/main.dart
@@ -1,49 +1,21 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:localsend_app/gen/strings.g.dart';
+import 'package:localsend_app/init.dart';
 import 'package:localsend_app/pages/home_page.dart';
 import 'package:localsend_app/provider/device_info_provider.dart';
 import 'package:localsend_app/provider/settings_provider.dart';
-import 'package:localsend_app/service/persistence_service.dart';
 import 'package:localsend_app/theme.dart';
 import 'package:localsend_app/util/device_info_helper.dart';
 import 'package:routerino/routerino.dart';
-import 'package:wechat_assets_picker/wechat_assets_picker.dart';
-import 'package:window_manager/window_manager.dart';
 
 Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-
-  if (!kIsWeb && defaultTargetPlatform == TargetPlatform.windows) {
-    await windowManager.ensureInitialized();
-    WindowManager.instance.setMinimumSize(const Size(400, 500));
-  }
-
-  if ([TargetPlatform.android, TargetPlatform.iOS, TargetPlatform.macOS].contains(defaultTargetPlatform)) {
-    try {
-      PhotoManager.clearFileCache();
-    } catch (e) {
-      print(e);
-    }
-  }
-
-  final persistenceService = await PersistenceService.initialize();
-
-  final locale = persistenceService.getLocale();
-  if (locale == null) {
-    LocaleSettings.useDeviceLocale();
-  } else {
-    LocaleSettings.setLocale(locale);
-  }
-
-  final deviceInfo = await getDeviceInfo();
-
+  final persistenceService = await preInit();
   runApp(TranslationProvider(
     child: ProviderScope(
       overrides: [
-        deviceRawInfoProvider.overrideWithValue(deviceInfo),
+        deviceRawInfoProvider.overrideWithValue(await getDeviceInfo()),
         settingsProvider.overrideWith((ref) => SettingsNotifier(persistenceService)),
       ],
       child: const LocalSendApp(),