+++ lib/provider/selected_files_provider.dart
@@ -1,3 +1,6 @@
+import 'dart:convert' show utf8;
+import 'dart:io';
+
 import 'package:file_picker/file_picker.dart' as file_picker;
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -5,7 +8,11 @@ import 'package:image_picker/image_picker.dart';
 import 'package:localsend_app/model/cross_file.dart';
 import 'package:localsend_app/model/file_type.dart';
 import 'package:localsend_app/util/file_path_helper.dart';
+import 'package:share_handler/share_handler.dart';
+import 'package:uuid/uuid.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
+
+const _uuid = Uuid();
 
 final selectedFilesProvider = StateNotifierProvider<SelectedFilesNotifier, List<CrossFile>>((ref) {
   return SelectedFilesNotifier();
@@ -13,6 +20,23 @@ final selectedFilesProvider = StateNotifierProvider<SelectedFilesNotifier, List<
 
 class SelectedFilesNotifier extends StateNotifier<List<CrossFile>> {
   SelectedFilesNotifier() : super([]);
+
+  /// Add a simple message
+  /// Internally, the message will be stored into [CrossFile.bytes] as UTF-8
+  void addMessage(String message) {
+    final List<int> bytes = utf8.encode(message);
+    state = List.unmodifiable([
+      ...state,
+      CrossFile(
+        name: '${_uuid.v4()}.txt',
+        fileType: FileType.text,
+        size: bytes.length,
+        asset: null,
+        path: null,
+        bytes: () async => bytes,
+      ),
+    ]);
+  }
 
   Future<void> addFiles<T>({
     required Iterable<T> files,
@@ -36,7 +60,7 @@ class SelectedFilesNotifier extends StateNotifier<List<CrossFile>> {
   }
 }
 
-/// Utility functions to convert third party models to [CrossFile].
+/// Utility functions to convert third party models to common [CrossFile] model.
 class CrossFileConverters {
   static Future<CrossFile> convertPlatformFile(file_picker.PlatformFile file) async {
     return CrossFile(
@@ -69,6 +93,19 @@ class CrossFileConverters {
       asset: null,
       path: kIsWeb ? null : file.path,
       bytes: kIsWeb ? () => file.readAsBytes() : null,
+    );
+  }
+
+  static Future<CrossFile> convertSharedAttachment(SharedAttachment attachment) async {
+    final file = File(attachment.path);
+    final fileName = attachment.path.fileName;
+    return CrossFile(
+      name: fileName,
+      fileType: fileName.guessFileType(),
+      size: await file.length(),
+      asset: null,
+      path: file.path,
+      bytes: null,
     );
   }
 }