+++ lib/widget/dialogs/message_input_dialog.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:localsend_app/gen/strings.g.dart';
+import 'package:routerino/routerino.dart';
+
+class MessageInputDialog extends StatefulWidget {
+  const MessageInputDialog();
+
+  @override
+  State<MessageInputDialog> createState() => _MessageInputDialogState();
+}
+
+class _MessageInputDialogState extends State<MessageInputDialog> {
+  String _text = '';
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text(t.dialogs.messageInput.title),
+      content: TextField(
+        keyboardType: TextInputType.multiline,
+        maxLines: null,
+        onChanged: (s) {
+          setState(() => _text = s);
+        },
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => context.pop(),
+          child: Text(t.general.cancel),
+        ),
+        ElevatedButton(
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Theme.of(context).buttonTheme.colorScheme!.primary,
+            foregroundColor: Theme.of(context).buttonTheme.colorScheme!.onPrimary,
+          ),
+          onPressed: () => context.pop(_text),
+          child: Text(t.general.confirm),
+        ),
+      ],
+    );
+  }
+}